+++ server/src/routes/api/receipts.js
@@ -1,15 +1,15 @@
 const router = require('express').Router()
-const { handleReceipt, handleTransactionHash } = require('@events/handlers')
+const { handleReceipt, handleTransactionHash } = require('@handlers/receipts')
 const { isEmpty } = require('@utils/web3/receipt')
 
 router.post('/', async (req, res, next) => {
-  const { receipt } = req.body
+  const { receipt, bridgeType } = req.body
   if (isEmpty(receipt)) {
     const msg = `Empty receipt given for transaction ${receipt.transactionHash}`
     console.log(msg)
     return res.json({ updated: false, msg })
   }
-  await handleReceipt(receipt)
+  await handleReceipt(receipt, bridgeType)
   return res.json({ updated: true })
 })
