+++ server/src/tasks/deployment/bridge.js
@@ -6,7 +6,7 @@ const IRestrictedTokenABI = require('@constants/abi/IRestrictedToken')
 const foreignAddressess = config.get('network.foreign.addresses')
 const homeAddresses = config.get('network.home.addresses')
 const { generateSignature } = require('@utils/web3')
-const { handleReceipt } = require('@events/handlers')
+const { handleReceipt } = require('@handlers/receipts')
 const mongoose = require('mongoose')
 const Token = mongoose.model('Token')
 const { setLengthLeft } = require('ethereumjs-util')
